+++ forecast/pipeline_multiple_replicas_forecast_tutorial/data/testdata_standard.df.json
@@ -1,1 +1,1 @@
-[{"cnt":[1526,1550,1708,1005,1623,1712,1530,1605,1538,1746,1472,1589,1913,1815,2115,2475,2927,1635,1812,1107,1450,1917,1807,1461,1969,2402,1446,1851]}]
+[{"count":[1526,1550,1708,1005,1623,1712,1530,1605,1538,1746,1472,1589,1913,1815,2115,2475,2927,1635,1812,1107,1450,1917,1807,1461,1969,2402,1446,1851]}]
