--- forecast/pipeline_multiple_replicas_forecast_tutorial/data/testdata_standard_dict.json
@@ -1,1 +1,0 @@
-"[{\"cnt\":1526},{\"cnt\":1550},{\"cnt\":1708},{\"cnt\":1005},{\"cnt\":1623},{\"cnt\":1712},{\"cnt\":1530},{\"cnt\":1605},{\"cnt\":1538},{\"cnt\":1746},{\"cnt\":1472},{\"cnt\":1589},{\"cnt\":1913},{\"cnt\":1815},{\"cnt\":2115},{\"cnt\":2475},{\"cnt\":2927},{\"cnt\":1635},{\"cnt\":1812},{\"cnt\":1107},{\"cnt\":1450},{\"cnt\":1917},{\"cnt\":1807},{\"cnt\":1461},{\"cnt\":1969},{\"cnt\":2402},{\"cnt\":1446},{\"cnt\":1851}]"
