+++ forecast/pipeline_multiple_replicas_forecast_tutorial/forecast-orchestration 2/.ipynb_checkpoints/main-checkpoint.py
@@ -1,0 +1,52 @@
+import wallaroo
+from wallaroo.object import EntityNotFoundError
+
+wl = wallaroo.Client()
+
+# get the arguments
+arguments = wl.task_args()
+
+if "workspace_name" in arguments:
+    workspace_name = arguments['workspace_name']
+else:
+    workspace_name="forecast-model-workshop"
+
+if "pipeline_name" in arguments:
+    pipeline_name = arguments['pipeline_name']
+else:
+    pipeline_name="bikedaypipe"
+
+def get_workspace(name):
+    workspace = None
+    for ws in wl.list_workspaces():
+        if ws.name() == name:
+            workspace= ws
+    return workspace
+
+def get_pipeline(name):
+    try:
+        pipeline = wl.pipelines_by_name(name)[0]
+    except EntityNotFoundError:
+        print(f"Pipeline not found:{name}")
+    return pipeline
+
+print(f"Workspace: {workspace_name}")
+workspace = get_workspace(workspace_name)
+
+wl.set_current_workspace(workspace)
+print(workspace)
+
+# the pipeline is assumed to be deployed
+print(f"Pipeline: {pipeline_name}")
+pipeline = get_pipeline(pipeline_name)
+print(pipeline)
+
+inferencedata = {"count": [1526, 1550, 1708, 1005, 1623, 1712, 1530, 1605, 1538, 1746, 1472, 1589, 1913, 1815, 2115, 2475, 2927, 1635, 1812, 1107, 1450, 1917, 1807, 1461, 1969, 2402, 1446, 1851]}
+
+print("Inference Data")
+print(inferencedata)
+
+print("Forecast")
+results = pipeline.infer(inferencedata)
+
+print(results)
