+++ forecast/pipeline_multiple_replicas_forecast_tutorial/resources/bigquery_service_account_statsmodel.json
@@ -1,16 +1,16 @@
 {
-    "type": "service_account",
-    "project_id": "YOUR PROJECT ID",
-    "private_key_id": "YOUR PRIVATE KEY ID",
+    "type": "YOUR TYPE",
+    "project_id": "YOUR PROJECT",
+    "private_key_id": "YOUR KEY",
     "private_key": "YOUR PRIVATE KEY",
     "client_email": "YOUR CLIENT EMAIL",
     "client_id": "YOUR CLIENT ID",
-    "auth_uri": "YOUR AUTH_URI",
-    "token_uri": "YOUR TOKEN_URI",
-    "auth_provider_x509_cert_url": "YOUR AUTH PROVIDER",
-    "client_x509_cert_url": "YOUR CLIENT X509 CERT URL",
-    "dataset":"YOUR DATASET",
+    "auth_uri": "YOUR AUTH URI",
+    "token_uri": "YOUR TOKEN URI",
+    "auth_provider_x509_cert_url": "YOUR AUTH PROVIDER URI",
+    "client_x509_cert_url": "YOUR CERT URL",
     "universe_domain": "YOUR UNIVERSE DOMAIN",
-    "input_table":"YOUR INPUTS TABLE",
-    "results_table":"YOUR RESULTS TABLE"
+    "dataset": "wallaroo_workshop_public",
+    "input_table": "bike_rentals",
+    "results_table": "bike_rentals_predictions"
 }
