--- forecast/pipeline_multiple_replicas_forecast_tutorial/resources/requirements.txt
@@ -1,3 +1,0 @@
-google-cloud-bigquery==3.10.0
-google-auth==2.17.3
-db-dtypes==1.1.1
