+++ forecast/pipeline_multiple_replicas_forecast_tutorial/step.py
@@ -2,5 +2,9 @@ import pandas as pd
 
 # takes the JSON output
 # column as `output`
-def wallaroo_json(data):
-    return pd.DataFrame({data})
+def wallaroo_json(data: pd.DataFrame):
+    # Add in code that averages forecast elements
+    return [
+             data,
+             {'average_rentals' : 1500}
+           ]
